Fix warning and remove unused codes

+++ drivers/leds/leds-microp.c
@@ -175,7 +175,7 @@ static void microp_led_jogball_brightness_set(struct led_classdev *led_cdev,enum
 	ldata->brightness = brightness;
 	spin_unlock_irqrestore(&ldata->brightness_lock, flags);
 
-	switch (brightness) {
+	switch ((int)brightness) {
 	case 0:
 		data[0] = 0;
 		break;
@@ -218,7 +218,7 @@ static void microp_led_wimax_brightness_set(struct led_classdev *led_cdev,enum l
 	ldata->brightness = brightness;
 	spin_unlock_irqrestore(&ldata->brightness_lock, flags);
 
-	switch (brightness) {
+	switch ((int)brightness) {
 	case 0:
 		data[0] = 0;
 		break;
